requirements.txt
omegaconf==2.3.0
Pillow==10.1.0
einops==0.7.0
git+https://github.com/tatsy/torchmcubes.git
transformers>=4.35.0
trimesh==4.0.5
rembg
huggingface-hub>=0.23.0
imageio[ffmpeg]
gradio
xatlas==0.0.9
moderngl==5.10.0
onnxruntime==1.20.1
fastapi==0.115.6
uvicorn==0.34.0
boto3
sagemaker
